check if Calibre is already installed

+++ macsetup.sh
@@ -2513,12 +2513,14 @@ installcask beardedspice
 # https://github.com/kovidgoyal/calibre
 # https://github.com/noDRM/DeDRM_tools
 #
-installcask calibre
-curl -L "$(lastversion noDRM/DeDRM_tools --assets)" --output Calibre-DeDRM.zip
-unzip Calibre-DeDRM.zip "DeDRM_plugin.zip"
-rm Calibre-DeDRM.zip
-notification "Install the DeDRM plugin into Calibre"
-rm DeDRM_plugin.zip
+if [[ ! $(isAppInstalled calibre) ]]; then
+    installcask calibre
+    curl -L "$(lastversion noDRM/DeDRM_tools --assets)" --output Calibre-DeDRM.zip
+    unzip Calibre-DeDRM.zip "DeDRM_plugin.zip"
+    rm Calibre-DeDRM.zip
+    notification "Install the DeDRM plugin into Calibre"
+    rm DeDRM_plugin.zip
+fi
 
 #
 # Captin
